+++ app/build.gradle.kts
@@ -1,0 +1,49 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+    alias(libs.plugins.ktor)
+    alias(libs.plugins.kotlin.plugin.serialization)
+}
+
+group = "com.eros"
+version = "0.0.1"
+
+application {
+    mainClass = "io.ktor.server.netty.EngineMain"
+}
+
+dependencies {
+    // Internal modules - all feature modules
+    implementation(project(":common"))
+    implementation(project(":database"))
+    implementation(project(":auth"))
+    implementation(project(":users"))
+    implementation(project(":wallet"))
+    implementation(project(":matching"))
+    implementation(project(":dates"))
+    implementation(project(":notifications"))
+
+    // Ktor server
+    implementation(libs.ktor.server.content.negotiation)
+    implementation(libs.ktor.server.core)
+    implementation(libs.ktor.serialization.kotlinx.json)
+    implementation(libs.ktor.server.rate.limiting)
+    implementation(libs.ktor.server.request.validation)
+    implementation(libs.ktor.server.default.headers)
+    implementation(libs.ktor.server.cors)
+    implementation(libs.ktor.server.call.logging)
+    implementation(libs.ktor.server.host.common)
+    implementation(libs.ktor.server.status.pages)
+    implementation(libs.ktor.server.auth)
+    implementation(libs.ktor.client.core)
+    implementation(libs.ktor.client.apache)
+    implementation(libs.ktor.server.auth.jwt)
+    implementation(libs.ktor.server.netty)
+    implementation(libs.ktor.server.config.yaml)
+
+    // Logging
+    implementation(libs.logback.classic)
+
+    // Testing
+    testImplementation(libs.ktor.server.test.host)
+    testImplementation(libs.kotlin.test.junit)
+}
