+++ auth/build.gradle.kts
@@ -1,0 +1,19 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+    alias(libs.plugins.kotlin.plugin.serialization)
+}
+
+dependencies {
+    // Internal modules
+    implementation(project(":common"))
+
+    // Ktor
+    implementation(libs.ktor.server.core)
+    implementation(libs.ktor.server.auth)
+    implementation(libs.ktor.server.auth.jwt)
+    implementation(libs.ktor.serialization.kotlinx.json)
+
+    // Testing
+    testImplementation(libs.kotlin.test.junit)
+    testImplementation(libs.ktor.server.test.host)
+}
