+++ dates/build.gradle.kts
@@ -1,0 +1,4 @@
+plugins {
+    alias(libs.plugins.kotlin.jvm)
+    alias(libs.plugins.kotlin.plugin.serialization)
+}
