+++ settings.gradle.kts
@@ -1,10 +1,13 @@
 rootProject.name = "eros-backend"
 
-include("common")
-
-dependencyResolutionManagement {
-    repositories {
-        mavenCentral()
-        maven("https://packages.confluent.io/maven/")
-    }
-}
+include(
+    "app",
+    "auth",
+    "common",
+    "database",
+    "dates",
+    "matching",
+    "notifications",
+    "users",
+    "wallet"
+)
