--- src/main/kotlin/com/eros/common/build.gradle.kts
@@ -1,13 +1,0 @@
-plugins {
-    id("application")
-}
-
-
-application {
-    mainClass = "com.eros.common.Main"
-}
-
-dependencies {
-    // Required for hashing passwords using Bcrypt.
-    implementation(libs.jbcrypt)
-}
